chore(entities-plugins/forms): retire former auto suggest
* Remove the former FieldAutoSuggest.
* Rename FieldAutoSuggestV2 to FieldAutoSuggest
